% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget.R
\name{gadget_binary}
\alias{gadget_binary}
\title{Return the location of the gadget binary installed by this package}
\usage{
gadget_binary()
}
\value{
File path, e.g. "/usr/local/lib/R/site-library/gadget/bin/gadget"
}
\description{
Return the location of the gadget binary installed by this package
}
\examples{
# Run gadget, capture help output
help_output <- system2(gadget_binary(), '-h', stdout = TRUE)
}
