\name{Rgadget-internal}
\alias{Rgadget-internal}
\alias{adjustconsumption}
\alias{catch}
\alias{derivedOptions}
\alias{distr}
\alias{eat}
\alias{firststep}
\alias{growthprob}
\alias{migrate}
\alias{recruits}
\alias{vonB}
\alias{suitability}
\docType{package}
\title{
Rgadget's internal functions not to be accessed by the user
}
\description{
  Internal functions in the Rgadget package.  
} 
\usage{
 nothing
}
\arguments{
 none
}
\details{
  none
}
\value{
  none
}
\keyword{internal}

